Refactor parser into SRP submodules
