+++ badges/ripr.json
@@ -1,6 +1,6 @@
 {
   "color": "orange",
   "label": "ripr",
-  "message": "12578",
+  "message": "12575",
   "schemaVersion": 1
 }
